==> code/04 cleaning.qmd <==
---
title: "04 cleaning"
format: html
editor: visual
---

# Setup
```{r}
library(tidyverse) 
library(readxl) 
library(openxlsx)
```


# Load data
must save data as csv in "02 fieldData.R" 
```{r}
df_csv <- read.csv("output/combined_data.csv")
na_percentage <- colMeans(is.na(df_csv)) \* 100 \# reading the csv from xlsx will change most values to logical
```

# Selecting out 100% NAs

```{r}
df_w <- df_csv %\>% 
  dplyr::select(all_of(na_percentage\[na_percentage == 100\])) %\>% 
  dplyr::select(-x20, -x22, -x23, -x24, -x25, -x26)
```
