==> code/07 cleaned length eda.qmd <==
---
title: "Untitled"
format: html
---

```{r}
library(tidyverse)
library(knitr)
```

```{r}
df <- read_csv("../output/cleaned/cleaned.cvs")
```

```{r}
df %>%
  summary
```

# EDA 
## Sites
```{r}
df %>%
  group_by(loc, loccode) %>%
  tally
```

## Years
```{r}
df %>%
  group_by(yr) %>%
  tally
```

## Site-years
How many site-years?
```{r}
df %>%
  group_by(loc, loccode, yr) %>%
  tally
```

```{r}
df %>%
  dplyr::select(loc, yr, rep, vcode, yt:e1, #elg, 
                mean_mic:mean_ui, -mean_uhm) %>%
  drop_na #%>%
  summary
  
  
  
```

uhml, mean_uhm, elg


```{r}
df %>%
  dplyr::select(uhml, mean_uhm, elg) %>%
  summary
  
  
df %>%
  dplyr::select(uhml, mean_uhm, elg) %>%
  filter(is.na(uhml) & is.na(mean_uhm)) #18,920 rowa

df %>%
  dplyr::select(uhml, mean_uhm, elg) %>%
  filter(!is.na(uhml) & !is.na(mean_uhm)) #0

df %>%
  dplyr::select(uhml, mean_uhm, elg) %>%
  filter(is.na(uhml) & !is.na(mean_uhm)) #9,963

df %>%
  dplyr::select(uhml, mean_uhm, elg) %>%
  filter(!is.na(uhml) & is.na(mean_uhm)) #10,073


nrow(df) - 20000

df %>%
  filter(uhml <= 5 | is.na(uhml)) %>%
  dplyr::select(uhml, mean_uhm) %>%
  mutate(mean_uhm = round(mean_uhm, 2)) %>%
  #summary
  pivot_longer(cols = everything()) %>%
  ggplot()+
  geom_density(aes(x=value, fill = name), alpha=.7)+
  #facet_wrap(~name, scales = "free")+
  labs(x = "Length value", y="Density",
       fill = "Length variable")+
  theme(legend.position = c(.15, .8))

ggsave("output/length_comparison.png",
       width = 6,
       height = 4)
```


```{r}
df %>% 
  dplyr::select(loc, loccode) %>%
  distinct() %>%
  arrange(loccode)
  filter(grepl("(*)", loc))
```

# Case when
```{r}
df %>%
  distinct(loc) %>%
  arrange(loc) %>%
  mutate(irrigation = case_when(
    grepl("iir|irri", loc) ~ "irrigated",
    grepl("dry", loc) ~ "dryland",
    T ~ NA
  )) %>%
  mutate(loc_fixed = case_when(
    grepl("Altus|ALTUS",loc) ~ "Altus, OK",
    grepl("Artesia", loc) ~ "Artesia, NM",
    T ~ loc
  ))


```

# Coordinates
```{r}
coords <- read_csv("../data/Cotton heatress Longitde and Latitude.csv") 

coords
```


```{r}
coords %>%
  mutate(loc_fixed = case_when(
    loccode == 19 & year == 97 ~ "College Station, TX",
    T ~ loc
  )) %>%
  filter(loccode == 19 & year == 97) %>%
  distinct()

grepl("pattern", colname)
```
