==> code/09 weather data daymetr.qmd <==
---
title: "09 weather data daymetr"
format: html
---

# Setup
```{r}
library(tidyverse)
library(daymetr)
library(readr)
library(dplyr)
library(lubridate)
library(ggplot2)
library(USAboundaries)
library(USAboundariesData)
library(reticulate)
```

```{r}
df <- read_csv("../data/daymet/index.csv")
```

```{r}
df_test <- df %>%
  filter(loc == "Altus, OK")

df_test %>%
  distinct(year) %>%
  nrow

30*365

df_test_w <- df %>%
  mutate(weather = pmap(list(.y = year,
                        .site = loc,
                        .lat = lat,
                        .lon = lon),
                        function(.y, .site, .lat, .lon)
                          download_daymet(
  site = .site,
  lat = .lat,
  lon = .lon,
  start = .y,
  end = .y,
  internal = TRUE,
  simplify = TRUE # return tidy data !!
  ) %>%
    rename(yr = year)
                          
                        
                        ))

df_unnest <- df_test_w %>%
  unnest(weather) %>%
  pivot_wider(names_from = measurement,
              values_from = value)


#data <- df_unnest
df_weather_unnest <- df_unnest
```

# write data table 
```{r data}
write.csv(df_weather_unnest, "data/daymet/unnestedWeatherData.csv")
tester <- read.csv("../data/daymet/unnestedWeatherData.csv")

data <- tester %>%
  select(-X) 

# temperature amplitude
data$tamp.deg.c. <- data$tmax..deg.c. - data$tmin..deg.c.
```

# GDD
```{r GDD}
#tmax
tb <- 15.6
data$gdd_tmax <- ifelse(data$tmax..deg.c. < 37.7 & data$tmax..deg.c. > tb, data$tmax..deg.c.,
                        ifelse(data$tmax..deg.c. >= 37.7, 37.7,
                               ifelse(data$tmax..deg.c. <= tb, tb, NA)))

ggplot(data = data, aes(x=tmax..deg.c., y=gdd_tmax))+
  geom_abline()+
  geom_point()
  
#tmin
data$gdd_tmin <- ifelse(data$tmin..deg.c. < 37.7 & data$tmin..deg.c. > tb, data$tmin..deg.c.,
                        ifelse(data$tmin..deg.c. >= 37.7, 37.7,
                               ifelse(data$tmin..deg.c. <= tb, tb, NA)))

ggplot(data = data, aes(x=tmin..deg.c., y=gdd_tmin))+
  geom_abline()+
  geom_point()

# tmax+tmin/2 - tb
data$gdd <- (data$gdd_tmax + data$gdd_tmin) / 2 - tb

data %>%
  dplyr::select(gdd_tmax, gdd_tmin, gdd)
               
```


# wrangling 
```{r}
# Create new columns for Year Start and Year End
data$year_Start <- as.Date(paste0(data$year, "-01-01"))
data$year_End <- as.Date(paste0(data$year, "-12-31"))

# dropped duplicated columns
df_final <- data %>%
select(-site, -lat, -lon, -yr) 


```


```{r growing season}
date <- as.Date(df_final$yday, origin = df_final$year_Start)
month <- as.numeric(format(date, "%m"))

# april/4 : nov:11 (growing season)
filtered_rows <- which(month>=4 & month<=11)

srad_mean_growingseason <- aggregate(df_final$srad..W.m.2.[filtered_rows], list(df_final$loc[filtered_rows], df_final$year[filtered_rows]), FUN = mean) %>% rename(loc = Group.1, year = Group.2, srad_mean_growingseason = x)

srad_sum_growingseason <- aggregate(df_final$srad..W.m.2.[filtered_rows], list(df_final$loc[filtered_rows], df_final$year[filtered_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>% 
    rename(srad_sum_growingseason = x)

tmax_mean_growingseason <- aggregate(df_final$tmax..deg.c.[filtered_rows], list(df_final$loc[filtered_rows], df_final$year[filtered_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(tmax_mean_growingseason = x)

tmin_mean_growingseason <- aggregate(df_final$tmin..deg.c.[filtered_rows], list(df_final$loc[filtered_rows], df_final$year[filtered_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(tmin_mean_growingseason = x)

vp_mean_growingseason <- aggregate(df_final$vp..Pa.[filtered_rows], list(df_final$loc[filtered_rows], df_final$year[filtered_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(vp_mean_growingseason = x)

dayl_mean_growingseason <- aggregate(df_final$dayl..s.[filtered_rows], list(df_final$loc[filtered_rows], df_final$year[filtered_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(dayl_mean_growingseason = x)

prcp_sum_growingseason <- aggregate(df_final$prcp..mm.day.[filtered_rows], list(df_final$loc[filtered_rows], df_final$year[filtered_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>%
    rename(prcp_sum_growingseason = x)

tamp_mean_growingseason <- aggregate(df_final$tamp.deg.c.[filtered_rows], list(df_final$loc[filtered_rows],
df_final$year[filtered_rows]), FUN = mean) %>%
select(-Group.1, -Group.2) %>%
rename(tamp_mean_growingseason = x)

gdd_mean_growingseason <- aggregate(df_final$gdd[filtered_rows], list(df_final$loc[filtered_rows],
df_final$year[filtered_rows]), FUN = mean) %>%
select(-Group.1, -Group.2) %>%
rename(gdd_mean_growingseason = x)

gdd_sum_growingseason <- aggregate(df_final$gdd[filtered_rows], list(df_final$loc[filtered_rows],
df_final$year[filtered_rows]), FUN = sum) %>%
select(-Group.1, -Group.2) %>%
rename(gdd_sum_growingseason = x)

# combined growingseasonmean 
growing_season_combined_data <- srad_mean_growingseason %>%
  bind_cols(srad_sum_growingseason) %>%
  bind_cols(tmax_mean_growingseason) %>%
  bind_cols(tmin_mean_growingseason) %>%
  bind_cols(tamp_mean_growingseason) %>%
  bind_cols(gdd_mean_growingseason) %>%
  bind_cols(gdd_sum_growingseason) %>%
  bind_cols(vp_mean_growingseason) %>%
  bind_cols(dayl_mean_growingseason) %>%
  bind_cols(prcp_sum_growingseason)
```

```{r every month}
months <- c(4:11)
monthly_combined <- growing_season_combined_data  # Create an initial data frame with columns from growing_season_combined_data

for (month in months) {
  month_rows <- which(as.numeric(format(date, "%m")) == month)
  
  srad_mean <- aggregate(df_final$srad..W.m.2.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>% 
    rename(!!paste0("srad_mean_", month) := x)
  
  assign(paste0("srad_mean_", month), srad_mean)

  srad_sum <- aggregate(df_final$srad..W.m.2.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>% 
    rename(!!paste0("srad_sum_", month) := x)

  assign(paste0("srad_sum_", month), srad_sum)
  
  tmax_mean <- aggregate(df_final$tmax..deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tmax_mean_", month) := x)
  
  assign(paste0("tmax_mean_", month), tmax_mean)

  tmin_mean <- aggregate(df_final$tmin..deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tmin_mean_", month) := x)
  
  assign(paste0("tmin_mean_", month), tmin_mean)

  vp_mean <- aggregate(df_final$vp..Pa.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("vp_mean_", month) := x)
  
  assign(paste0("vp_mean_", month), vp_mean)

  dayl_mean <- aggregate(df_final$dayl..s.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("dayl_mean_", month) := x)
  
  assign(paste0("dayl_mean_", month), dayl_mean)

  prcp_sum <- aggregate(df_final$prcp..mm.day.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("prcp_sum_", month) := x)
  
  assign(paste0("prcp_sum_", month), prcp_sum)

  tamp_mean <- aggregate(df_final$tamp.deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tamp_mean_", month) := x)
  
  assign(paste0("tamp_mean_", month), tamp_mean)

 gdd_mean <- aggregate(df_final$gdd[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("gdd_mean_", month) := x)
  
  assign(paste0("gdd_mean_", month), gdd_mean)

 gdd_sum <- aggregate(df_final$gdd[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("gdd_sum_", month) := x)
  
  assign(paste0("gdd_sum_", month), gdd_sum)
  
  # Combine the current month's data frames with the previously combined data
  current_month_combined <- bind_cols(get(paste0("srad_mean_", month)), 
                                      get(paste0("srad_sum_", month)), 
                                      get(paste0("tmax_mean_", month)), 
                                      get(paste0("tmin_mean_", month)),
                                      get(paste0("tamp_mean_", month)),
                                      get(paste0("gdd_mean_", month)),
                                      get(paste0("gdd_sum_", month)),
                                      get(paste0("vp_mean_", month)), 
                                      get(paste0("dayl_mean_", month)), 
                                      get(paste0("prcp_sum_", month)))
  
  monthly_combined <- bind_cols(monthly_combined, current_month_combined)

 #rename numbers to month names April-Nov (4-11)
  monthly_combined <- monthly_combined %>%
  rename_with(~sub("4", "April", .), contains("4")) %>%
  rename_with(~sub("5", "May", .), contains("5")) %>%
  rename_with(~sub("6", "June", .), contains("6")) %>%
  rename_with(~sub("7", "July", .), contains("7")) %>%
  rename_with(~sub("8", "August", .), contains("8")) %>%
  rename_with(~sub("9", "September", .), contains("9")) %>%
  rename_with(~sub("10", "October", .), contains("10")) %>%
  rename_with(~sub("11", "November", .), contains("11")) 
}

```

```{r every 2 months}
months <- c(4, 6, 8, 10)
twice_combined <- monthly_combined  # Create an initial data frame with columns from growing_season_combined_data

for (i in seq_along(months)) {
  month1 <- months[i]
  month2 <- months[i] + 1
  
  month_rows <- which(as.numeric(format(date, "%m")) %in% c(month1, month2))
  
  srad_mean2 <- aggregate(df_final$srad..W.m.2.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>% 
    rename(!!paste0("srad_mean_", month1, month2) := x)
  
  assign(paste0("srad_mean_", month1, month2), srad_mean2)

  srad_sum2 <- aggregate(df_final$srad..W.m.2.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN =sum) %>%
    select(-Group.1, -Group.2) %>% 
    rename(!!paste0("srad_sum_", month1, month2) := x)

  assign(paste0("srad_sum_", month1, month2), srad_sum2)
  
  tmax_mean2 <- aggregate(df_final$tmax..deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tmax_mean_", month1, month2) := x)
  
  assign(paste0("tmax_mean_", month1, month2), tmax_mean2)

  tmin_mean2 <- aggregate(df_final$tmin..deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tmin_mean_", month1, month2) := x)
  
  assign(paste0("tmin_mean_", month1, month2), tmin_mean2)

  vp_mean2 <- aggregate(df_final$vp..Pa.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("vp_mean_", month1, month2) := x)
  
  assign(paste0("vp_mean_", month1, month2), vp_mean2)

dayl_mean2 <- aggregate(df_final$dayl..s.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("dayl_mean_", month1, month2) := x)
  
  assign(paste0("dayl_mean_", month1, month2), dayl_mean2)

  prcp_sum2 <- aggregate(df_final$prcp..mm.day.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("prcp_sum_", month1, month2) := x)
  
  assign(paste0("prcp_sum_", month1, month2), prcp_sum2)

  tamp_mean2 <- aggregate(df_final$tamp.deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tamp_mean_", month1, month2) := x)
  
  assign(paste0("tamp_mean_", month1, month2), tamp_mean2)

  gdd_mean2 <- aggregate(df_final$gdd[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("gdd_mean_", month1, month2) := x)
  
  assign(paste0("gdd_mean_", month1, month2), gdd_mean2)

  gdd_sum2 <- aggregate(df_final$gdd[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("gdd_sum_", month1, month2) := x)
  
  assign(paste0("gdd_sum_", month1, month2), gdd_sum2)
  
  # Combine the current month's data frames with the previously combined data
  current_twice_combined <- bind_cols(get(paste0("srad_mean_", month1, month2)), 
                                      get(paste0("srad_sum_", month1, month2)), 
                                      get(paste0("tmax_mean_", month1, month2)), 
                                      get(paste0("tmin_mean_", month1, month2)),
                                      get(paste0("tamp_mean_", month1, month2)),
                                      get(paste0("gdd_mean_", month1, month2)),
                                      get(paste0("gdd_sum_", month1, month2)),
                                      get(paste0("vp_mean_", month1, month2)),
                                      get(paste0("dayl_mean_", month1, month2)), 
                                      get(paste0("prcp_sum_", month1, month2)))
  
  twice_combined <- bind_cols(twice_combined, current_twice_combined)
  
  # Rename the columns with the combined month values
  twice_combined <- twice_combined %>%
    rename_with(~sub(paste0(month1, month2), paste0(month1, "_", month2), .)) %>%
    rename_with(~sub("4", "April", .), contains("4")) %>%
    rename_with(~sub("5", "May", .), contains("5")) %>%
    rename_with(~sub("6", "June", .), contains("6")) %>%
    rename_with(~sub("7", "July", .), contains("7")) %>%
    rename_with(~sub("8", "August", .), contains("8")) %>%
    rename_with(~sub("9", "September", .), contains("9")) %>%
    rename_with(~sub("10", "October", .), contains("10")) %>%
    rename_with(~sub("11", "November", .), contains("11")) 
  }

```

```{r every 4 months}
months <- c(4, 8)
four_combined <- twice_combined  # Create an initial data frame with columns from growing_season_combined_data

for (i in seq_along(months)) {
  month1 <- months[i]
  month2 <- months[i] + 1
  month3 <- months[i] + 2
  month4 <- months[i] + 3
  
  month_rows <- which(as.numeric(format(date, "%m")) %in% c(month1, month2, month3, month4))
  
  srad_mean4 <- aggregate(df_final$srad..W.m.2.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>% 
    rename(!!paste0("srad_mean_", month1, month2, month3, month4) := x)
  
  assign(paste0("srad_mean_", month1, month2, month3, month4), srad_mean4)

  srad_sum4 <- aggregate(df_final$srad..W.m.2.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>% 
    rename(!!paste0("srad_sum_", month1, month2, month3, month4) := x)

  assign(paste0("srad_sum_", month1, month2, month3, month4), srad_sum4)
  
  tmax_mean4 <- aggregate(df_final$tmax..deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tmax_mean_", month1, month2, month3, month4) := x)
  
  assign(paste0("tmax_mean_", month1, month2, month3, month4), tmax_mean4)

  tmin_mean4 <- aggregate(df_final$tmin..deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tmin_mean_", month1, month2, month3, month4) := x)
  
  assign(paste0("tmin_mean_", month1, month2, month3, month4), tmin_mean4)

  vp_mean4 <- aggregate(df_final$vp..Pa.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("vp_mean_", month1, month2, month3, month4) := x)
  
  assign(paste0("vp_mean_", month1, month2, month3, month4), vp_mean4)

  dayl_mean4 <- aggregate(df_final$dayl..s.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("dayl_mean_", month1, month2, month3, month4) := x)
  
  assign(paste0("dayl_mean_", month1, month2, month3, month4), dayl_mean4)

  prcp_sum4 <- aggregate(df_final$prcp..mm.day.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("prcp_sum_", month1, month2, month3, month4) := x)
  
  assign(paste0("prcp_sum_", month1, month2, month3, month4), prcp_sum4)

  tamp_mean4 <- aggregate(df_final$tamp.deg.c.[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("tamp_mean_", month1, month2, month3, month4) := x)
  
  assign(paste0("tamp_mean_", month1, month2, month3, month4), tamp_mean4)

  gdd_mean4 <- aggregate(df_final$gdd[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = mean) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("gdd_mean_", month1, month2, month3, month4) := x)
  
  assign(paste0("gdd_mean_", month1, month2, month3, month4), gdd_mean4)

  gdd_sum4 <- aggregate(df_final$gdd[month_rows], list(df_final$loc[month_rows], df_final$year[month_rows]), FUN = sum) %>%
    select(-Group.1, -Group.2) %>%
    rename(!!paste0("gdd_sum_", month1, month2, month3, month4) := x)
  
  assign(paste0("gdd_sum_", month1, month2, month3, month4), gdd_sum4)
  
  # Combine the current month's data frames with the previously combined data
  current_four_combined <- bind_cols(get(paste0("srad_mean_", month1, month2, month3, month4)), 
                                      get(paste0("srad_sum_", month1, month2, month3, month4)), 
                                      get(paste0("tmax_mean_", month1, month2, month3, month4)), 
                                      get(paste0("tmin_mean_", month1, month2, month3, month4)),
                                      get(paste0("tamp_mean_", month1, month2, month3, month4)),
                                      get(paste0("gdd_mean_", month1, month2, month3, month4)),
                                      get(paste0("gdd_sum_", month1, month2, month3, month4)),
                                      get(paste0("vp_mean_", month1, month2, month3, month4)), 
                                      get(paste0("dayl_mean_", month1, month2, month3, month4)), 
                                      get(paste0("prcp_sum_", month1, month2, month3, month4)))
  
  four_combined <- bind_cols(four_combined, current_four_combined)
  
  # Rename the columns with the combined month values
  four_combined <- four_combined %>%
    rename_with(~sub(paste0(month1, month2, month3, month4), paste0(month1, "_", month2, "_", month3, "_", month4), .)) %>%
    rename_with(~sub("4", "April", .), contains("4")) %>%
    rename_with(~sub("5", "May", .), contains("5")) %>%
    rename_with(~sub("6", "June", .), contains("6")) %>%
    rename_with(~sub("7", "July", .), contains("7")) %>%
    rename_with(~sub("8", "August", .), contains("8")) %>%
    rename_with(~sub("9", "September", .), contains("9")) %>%
    rename_with(~sub("10", "October", .), contains("10")) %>%
    rename_with(~sub("11", "November", .), contains("11")) 
  }

```

twice_combined has
1. growing_season
2. monthly (april-nov)
3. every two months (april-may, june-july, aug-sep, oct-nov)

four_combined (Final data frame)
1. growing season
2. monthly (april-nov)
3. every two months (april-may, june-july, aug-sep, oct-nov)
4. every 4 months (April, may, june, july) & (Aug, sep, oct, nov)


```{r four_combined}
write.csv(four_combined, "data/daymet/weatherDataDaymetrFinal.csv")

```


# plotting

```{r}
library(ggplot2)
```


```{r density plots}
# growing season VP 
densityplot <- ggplot(twice_combined, aes(x=vp_mean_growingseason)) + 
  geom_density()

densityplot
# Add mean line
densityplot+ geom_vline(aes(xintercept=mean(vp_mean_growingseason)),
            color="blue", linetype="dashed", size=1)
```
