==> code/11 combine field and weather.qmd <==
---
title: "11 combine field and weather"
format: html
---

# Setup
```{r}
library(tidyverse)
```


```{r weather}
weather <- read_csv("../data/daymet/weatherDataDaymetrFinal.csv")

weather
```

```{r}
str_blue <- read_csv("../data/daymet/field_blues_str.csv")

str_blue
```

# Joining 
```{r}
fieldweather <- str_blue %>%
  left_join(weather) %>%
  dplyr::select(-`...1`) 

fieldweather
```

```{r}
write_csv(fieldweather,
          "../data/daymet/fieldweather_final.csv")

```
