==> code/12 exploratory regression.qmd <==
Run some exploratory regressions with fiber strength vs. year

overall (one regression line)
per region (SE, SW, SC, etc.), one line per region
per state, one line per state

# setup 

```{r}
library(tidyverse)
library(readr)
library(ggplot2)
library(dplyr)
library(xlsx)
library(writexl)
```

# load data
```{r}
tester <- read.csv("data/daymet/fieldweather_final.csv")
```

# wrangling data / removing alaska point 
```{r}
data <- tester %>%
filter(loc != "Palmer, AK")
```

```{r overall}
# linear regression model
model <- lm(str_emmean ~ year, data = data)

# scatter plot w regression line for overall 
# fiber strength vs year 
overallPlot <- ggplot(data, aes(x = year, y = str_emmean)) +
  geom_point() +
  geom_smooth(method = "lm", se = FALSE, color = "red") +
  labs(x = "Year", y = "Fiber Strength")

ggsave("data/regression/overallPlot.png", overallPlot)
```

```{r per region} 
# Far West, Southwest, Midsouth, Southeast
# https://www.cottoninc.com/cotton-production/quality/us-cotton-fiber-chart/
data <- data %>%
  mutate(region = case_when(
    grepl("CA|AZ|NM|NV", loc) ~ "FarWest",
    grepl("TX|OK|KS", loc) ~ "Southwest",
    grepl("MO|AR|LA|MS|TN", loc) ~ "Midsouth",
    grepl("AL|GA|SC|NC|VA|FL", loc)  ~ "Southeast",
    TRUE ~ "Other"
  ))


# shows grapht plots
region_plot <- data %>%
  ggplot(aes(x = year, y = str_emmean, color = region)) +
    geom_point() +
    geom_smooth(method = "lm", se = FALSE)+
  facet_grid(.~region)+
  theme(legend.position = "none")

ggsave("data/regression/region_plot.png", region_plot)
# 
str ~ region*year
str ~ region + year + region:year


```


```{r per state}
data <- data %>%
  mutate(state = case_when(
    grepl("TX", loc) ~ "Texas",
    grepl("OK", loc) ~ "Oklahoma",
    grepl("KS", loc) ~ "Kansas",
    grepl("MO", loc) ~ "Missouri",
    grepl("AR", loc) ~ "Arkansas",
    grepl("LA", loc) ~ "Louisiana",
    grepl("MS", loc) ~ "Mississippi",
    grepl("TN", loc) ~ "Tennessee",
    grepl("AL", loc) ~ "Alabama",
    grepl("GA", loc) ~ "Georgia",
    grepl("SC", loc) ~ "South Carolina",
    grepl("NC", loc) ~ "North Carolina",
    grepl("VA", loc) ~ "Virginia",
    grepl("FL", loc) ~ "Florida",
    grepl("CA", loc) ~ "California",
    grepl("AZ", loc)  ~ "Arizona",
    grepl("NM", loc ) ~ "New Mexico",
    grepl("NV", loc) ~ "Nevada",
    TRUE ~ "Other"
  ))

state_plot <- data %>%
    filter(loc != "Logandale, NV") %>% # logandale only has one point (not useful in this information)
  ggplot(aes(x = year, y = str_emmean, color = state)) +
    geom_point() +
    geom_smooth(method = "lm", se = FALSE)+
  facet_grid(.~state)+
  theme(legend.position = "none")

  ggsave("data/regression/state_plot.png", state_plot)

str ~ region*year
str ~ region + year + region:year

```

# save region / state data 
```{r}
write_xlsx(data, "data/daymet/finalWeatherandRegions.xlsx")
write_csv(data, "data/daymet/finalWeatherandRegions.csv")
```
