==> code/16 plots.qmd <==
# setup 
```{r}
library(tidyverse)
library(readxl)
library(janitor)
library(sf)
library(stringr)
library(readr)
library(dplyr)
library(ggplot2)
library(USAboundaries)
library(USAboundariesData)
```

# load data 
```{r}
coords <- read_excel("../data/Cotton heatress Longitde and Latitude.xlsx")
```


# Wrangling ########################################

```{r}
coords_w <- coords %>%
  clean_names() %>%
  # Select target columns  
  dplyr::select(year, reg_code, loc, loccode, varcode, latitude, longitude) %>%
  filter(loc != "Palmer, AZ") %>%
  filter(loc != "?") %>%
  mutate(latitude = gsub("° N", "", as.character(latitude)),
         longitude = gsub("° W", "", as.character(longitude))) %>%
#         loc = gsub("AZ", "AK", as.character(loc))) %>%
  mutate(latitude = as.numeric(latitude),
         longitude = as.numeric(longitude))

## View entire decimal value  ########
options(digits =6)
print(as.data.frame(coords_w[30:45, ]))

## Transform into geospatial object #################
# remove NA and * -1 
test <- coords_w %>%
  drop_na(latitude, longitude) %>%
  mutate(longitude = longitude *-1)

# EDA ###############
# Will check data with tabular summaries and graphs. - `summary()` - `ggplot()`
## Plotting on state boundaries graph ###########
states_contemporary <- us_states() %>%
    filter(!state_abbr %in% c("AK", "HI", "PR"))

location_counts <- test %>%
  group_by(loc, latitude, longitude) %>%
  summarize(count = n())

coord_plot <- ggplot() +
    geom_sf(data = states_contemporary, fill = NA) +
    geom_point(data = location_counts, aes(x = longitude, y = latitude, color = count), size = 3) +
    scale_color_gradient(low = "lightgreen", high = "darkblue") +
    xlim(-130, -60) + 
    labs(title = "Site Locations") + 
    theme(plot.title = element_text(size = 20, hjust = 0.5))


coord_plot

```


# find highest blues
```{r}
blues <- read.csv("../data/daymet/field_blues_str.csv")

max_values <- blues %>%
  filter(str_emmean == max(str_emmean)) %>%
  select(year, loc, lat, lon, str_emmean) %>%
  mutate(lon = lon *-1) %>%
  distinct()

min_values <- blues %>%
  filter(str_emmean == min(str_emmean)) %>%
  select(year, loc, lat, lon, str_emmean) %>%
  mutate(lon = lon *-1) %>%
  distinct()
```

# edit map 
```{r}
coord_plot <- ggplot() +
  geom_sf(data = states_contemporary, fill = "gray80", color = "black")+
      geom_point(data = max_values, aes(x = lon, y = lat), fill = "red", shape = 21, size = 10, alpha = .5) +
  geom_point(data = min_values, aes(x = lon, y = lat), fill = "blue", shape = 21, size = 10, alpha =  .5) +
  geom_point(data = location_counts, 
             aes(x = longitude, y = latitude, 
                 fill = count), shape=21, color = "black",
             size = 3) +
  #geom_point(data = max_values, aes(x = lon, y = lat), color = "red", shape = 8, size = 10, alpha = 2) +
  #geom_point(data = min_values, aes(x = lon, y = lat), color = "blue", shape = 8, size = 10, alpha =  2) +
  #scale_fill_viridis_b()+
  scale_fill_gradient(low = "lightgreen", 
                      high = "darkblue") +
  xlim(-123, -68) + 
  labs(title = "Cotton Site Locations") + 
      ggthemes::theme_map()+
  theme(plot.title = element_text(size = 35, hjust = 0.5),
        #axis.title = element_text(size = 35),
        #axis.text = element_text(size = 30),
        legend.direction = "horizontal",
        legend.text = element_text(size = 16),
        legend.title = element_text(size = 18),
        #legend.position = c(0.95, 0.4),
        legend.key.width = unit(2, 'cm'),
        legend.position = c(.9, 0.05),
        legend.justification = c(1, 1))


coord_plot

ggsave("../figs/updatedMap.png", 
       plot = coord_plot,
       bg = "white",
       width = 6,
       height = 5) #, width = 8, height = 6, dpi = 300)
```

# Daily
