==> code/17 weather plots.qmd <==
# setup 
```{r}
library(dplyr)
library(tidyverse)
library(patchwork)
```

```{r}
tester <- read.csv("data/daymet/unnestedWeatherData.csv")
```

# secondary
#tamp / gdd 
```{r}
tester$tamp.deg.c. <- tester$tmax..deg.c. - tester$tmin..deg.c.

#tmax
tb <- 15.6
tester$gdd_tmax <- ifelse(tester$tmax..deg.c. < 37.7 & tester$tmax..deg.c. > tb, tester$tmax..deg.c.,
                        ifelse(tester$tmax..deg.c. >= 37.7, 37.7,
                               ifelse(tester$tmax..deg.c. <= tb, tb, NA)))

#tmin
tester$gdd_tmin <- ifelse(tester$tmin..deg.c. < 37.7 & tester$tmin..deg.c. > tb, tester$tmin..deg.c.,
                        ifelse(tester$tmin..deg.c. >= 37.7, 37.7,
                               ifelse(tester$tmin..deg.c. <= tb, tb, NA)))

# tmax+tmin/2 - tb
tester$gdd <- (tester$gdd_tmax + tester$gdd_tmin) / 2 - tb
```

# plot
```{r}
weather_1980 <- subset(tester, year == 1980)

weather_2020 <- subset(tester, year == 2020)

combined_data <- bind_rows(
  weather_1980 %>% mutate(Year = "1980"),
  weather_2020 %>% mutate(Year = "2020")
)

# Create a new ggplot using the combined data
tmax_combined_plot <- ggplot(combined_data, aes(x = tmax..deg.c., fill = Year)) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(values = c("#008000", "#FFA500")) + 
  labs(title = "Tmax in 1980 and 2020", x = "tmax..deg.c.") + 
   ylab(NULL) + 
  theme(
    plot.title = element_text(size = 35, hjust = 0.5),
     legend.position = "none",
    axis.title = element_text(size = 30),
    axis.text = element_text(size = 30),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank() 
  )

 tmin_combined_plot <- ggplot(combined_data, aes(x = tmin..deg.c., fill = Year)) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(values = c("#008000", "#FFA500")) + 
 labs(title = "Tmin in 1980 and 2020") +
  ylab(NULL) + 
  theme(
    plot.title = element_text(size = 35, hjust = 0.5),
    axis.title = element_text(size = 30),
    axis.text = element_text(size = 30),
    legend.text = element_text(size = 30),
    legend.title = element_text(size = 30),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank() 
  )


primary_combined <- tmax_combined_plot + tmin_combined_plot + plot_layout(ncol = 2)


ggsave("figs/tmax_tmin.png", plot = primary_combined, height =7, width = 14)
```

```{r}

tamp_combined_plot <- ggplot(combined_data, aes(x = tamp.deg.c., fill = Year)) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(values = c("#008000", "#FFA500")) + 
  labs(title = "Tamp in 1980 and 2020", x = "tamp.deg.c.") + 
  ylab(NULL) + 
  theme(
    plot.title = element_text(size = 35, hjust = 0.5),
      legend.position = "none",
    axis.title = element_text(size = 30),
    axis.text = element_text(size = 30),
    legend.text = element_text(size = 30),
    legend.title = element_text(size = 30),
     axis.text.y = element_blank(),
    axis.ticks.y = element_blank() 
  )
gdd_combined_plot <- ggplot(combined_data, aes(x = gdd, fill = Year)) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(values = c("#008000", "#FFA500")) + 
  labs(title = "GDD in 1980 and 2020", x = "gdd") + 
  ylab(NULL) + 
  theme(
    plot.title = element_text(size = 35, hjust = 0.5),
    axis.title = element_text(size = 30),
    axis.text = element_text(size = 30),
    legend.text = element_text(size = 30),
    legend.title = element_text(size = 30),
     axis.text.y = element_blank(),
    axis.ticks.y = element_blank() 
  )

secondary_combined <- tamp_combined_plot + gdd_combined_plot + plot_layout(ncol = 2)


ggsave("figs/tamp_gdd.png", plot = secondary_combined, height =7, width = 14)

```

# blues 
```{r}
blues <- read.csv("data/daymet/field_blues_str.csv")

max_values <- blues %>%
  filter(str_emmean == max(str_emmean)) %>%
  select(year, loc, str_emmean) %>%
  distinct()

min_values <- blues %>%
  filter(str_emmean == min(str_emmean)) %>%
  select(year, loc, str_emmean) %>%
  distinct()

```

# summarized
```{r}
summarized <- read.csv("../data/daymet/finalWeatherandRegions.csv") 

growingseason <- summarized %>%
  dplyr::select(tamp_mean_growingseason, str_emmean, loc, year)

# plot 
max_locations <- summarized %>% # red star on map 
  filter(loc %in% max_values$loc)

min_locations <- summarized %>% # blue star on map
  filter(loc %in% min_values$loc)

density_plot <- ggplot() +
  geom_density(data = max_locations, aes(x = tamp_mean_growingseason, fill = "Max"), alpha = 0.5) +
  geom_density(data = min_locations, aes(x = tamp_mean_growingseason, fill = "Min"), alpha = 0.5) +
  scale_fill_manual(values = c("red", "blue"),
                    labels = c("Max", "Min"),
                    name = "str_emmean") +
  labs(title = "Tamp Plot - Growing Season", x = "Temperature Amplitude", y = "Density") +
  theme(plot.title = element_text(size = 40, hjust = 0.5),
        axis.title = element_text(size = 30),
        axis.text = element_text(size = 30),
        legend.text = element_text(size = 30),
        legend.title = element_text(size = 30),
        legend.position = c(0.8, 0.8),         
        legend.justification = c(1, 1),
        plot.background = element_rect(fill = "white", color = "black"),
        plot.margin = margin(4, 2, 4, 3, "cm"))        

density_plot

ggsave("figs/summarized_growingseason.png", plot = density_plot) # width = 8, height = 6, dpi = 300)

# work on later ## 
quadra_months <- summarized %>%
  dplyr::select(c(tamp_mean_April_May_June_July,tamp_mean_August_September_October_November), str_emmean, loc, year)

quadra_plot <- ggplot() +
  geom_density(data = max_locations, aes(x = quadra_months, fill = "Max"), alpha = 0.5) +
  geom_density(data = min_locations, aes(x = quadra_months, fill = "Min"), alpha = 0.5) +
  scale_fill_manual(values = c("red", "blue"),
                    labels = c("Max", "Min"),
                    name = "str_emmean") +
  labs(title = "Density Plot - Quadra Months", x = "Quadra Months", y = "Density") +
  theme(plot.title = element_text(size = 20, hjust = 0.5),
        axis.title = element_text(size = 16),
        axis.text = element_text(size = 14),
        legend.text = element_text(size = 14),
        legend.title = element_text(size = 16),
        legend.position = c(0.8, 0.8),
        legend.justification = c(1, 1))




bimonths <- summarized %>%
  dplyr::select(c(tamp_mean_April_May, tamp_mean_October_November), str_emmean, loc, year)

months <- summarized %>%
  dplyr::select(c(tamp_mean_April,tamp_mean_November,tamp_mean_May, tamp_mean_July, tamp_mean_August, tamp_mean_September, tamp_mean_June), str_emmean, loc, year)

```

```{r}
summarized %>%
  filter(loc %in% c("Keiser, AR", "Yuma, AZ")) %>%
  dplyr::select(contains("tamp"))
```


# Daily tamp plot  
```{r}
daily_tamp <- tester %>%
  filter(loc == "Keiser, AR") %>%
  filter(year == 1992)

timescales <- tribble(~scale, ~begin, ~end,
                      "Growing season", 91, 334,
                      "Every 4 months", 91, 212,
                      "Every 2 months", 91, 151,
                      "Monthly", 91, 120
                      ) %>%
  mutate(scale = factor(scale,
                        levels = c("Monthly",
                                   "Every 2 months",
                                   "Every 4 months",
                                   "Growing season")
                        ))

timescales_plot <- timescales %>%
  ggplot(aes(y = fct_rev(scale)))+
  geom_linerange(aes(color = scale, xmin = begin, xmax = end),
                 linewidth = 2)+
  labs(x = "Day of year",
       y = "Scale")+
  scale_x_continuous(breaks = seq(0,365, 73),
                    limits = c(0,365))+
  ggthemes::scale_color_colorblind()+
  theme(legend.position = "none",
        panel.grid.minor = element_blank(),
        text = element_text(size = 20))

timescales_plot

daily_tamp_plot <- daily_tamp  %>%
  ggplot(aes(x=yday, y=tamp.deg.c.))+
  geom_area(color = "black", size = .8, fill = "purple", alpha = .7)+
  #geom_line()+
  #geom_point()+
    scale_x_continuous(breaks = seq(0,365, 73),
                    limits = c(0,365))+
  labs(x="Day of year")+
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        panel.grid.minor = element_blank(),
        text = element_text(size = 20))
  
library(patchwork)
daily_tamp_plot / timescales_plot +
  plot_layout(heights = c(4,1))

ggsave("../figs/timescales.png",
       width = 10,
       height = 5)
```
